response 구조 수정
